Separate ollama_llm_cost.py
